requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
gymnasium>=0.29.0
yfinance>=0.2.35
plotly>=5.18.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark==4.0.0
dash==3.0.2
ta>=0.10.2
mypy>=1.8.0
pylint>=3.0.0
black>=24.1.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
types-requests>=2.31.0
types-python-dateutil>=2.8.19
types-Flask>=1.1.6
types-Werkzeug>=1.0.9
types-Jinja2>=2.11.9
pandas-stubs>=2.0.0
types-seaborn>=0.13.0